--- a/diagnose.cabal
+++ b/diagnose.cabal
@@ -65,7 +65,8 @@ library
       BlockArguments
   ghc-options: -Wall -Wextra
   build-depends:
-      base >=4.7 && <5
+      array ==0.5.*
+    , base >=4.7 && <5
     , data-default >=0.7 && <1
     , dlist ==1.0.*
     , hashable >=1.3 && <2
@@ -109,7 +110,8 @@ test-suite diagnose-megaparsec-tests
       BlockArguments
   ghc-options: -Wall -Wextra -threaded -rtsopts -with-rtsopts=-N
   build-depends:
-      base >=4.7 && <5
+      array ==0.5.*
+    , base >=4.7 && <5
     , data-default >=0.7 && <1
     , diagnose
     , dlist ==1.0.*
@@ -149,7 +151,8 @@ test-suite diagnose-parsec-tests
       BlockArguments
   ghc-options: -Wall -Wextra -threaded -rtsopts -with-rtsopts=-N
   build-depends:
-      base >=4.7 && <5
+      array ==0.5.*
+    , base >=4.7 && <5
     , data-default >=0.7 && <1
     , diagnose
     , dlist ==1.0.*
@@ -188,7 +191,8 @@ test-suite diagnose-rendering-tests
       BlockArguments
   ghc-options: -Wall -Wextra -threaded -rtsopts -with-rtsopts=-N
   build-depends:
-      base >=4.7 && <5
+      array ==0.5.*
+    , base >=4.7 && <5
     , data-default >=0.7 && <1
     , diagnose
     , dlist ==1.0.*