--- a/package.yaml
+++ b/package.yaml
@@ -8,11 +8,12 @@ category: "Error Reporting"
 
 dependencies:
 - base >= 4.7 && < 5
+- data-default >= 0.7 && < 1
+- dlist >= 1.0 && < 1.1
+- hashable >= 1.3 && < 2
 - prettyprinter >= 1.7.0 && < 2
 - prettyprinter-ansi-terminal >= 1.1.0 && < 2
 - unordered-containers >= 0.2 && < 0.3
-- hashable >= 1.3 && < 2
-- data-default >= 0.7 && < 1
 - wcwidth >= 0.0.1 && <1
 - text >= 1.0.0.0 && <= 2.0
 # ^^^ This is unfortunately required, but as 'prettyprinter' already depends on it, it will already have been fetched