--- a/package.yaml
+++ b/package.yaml
@@ -8,6 +8,7 @@ category: "Error Reporting"
 
 dependencies:
 - base >= 4.7 && < 5
+- array >= 0.5 && < 0.6
 - data-default >= 0.7 && < 1
 - dlist >= 1.0 && < 1.1
 - hashable >= 1.3 && < 2