--- a/src/Error/Diagnose/Diagnostic/Internal.hs
+++ b/src/Error/Diagnose/Diagnostic/Internal.hs
@@ -21,15 +21,15 @@ import Data.Aeson (ToJSON(..), encode, object, (.=))
 import Data.ByteString.Lazy (ByteString)
 #endif
 
+import Data.Array (listArray)
 import Data.DList (DList)
 import qualified Data.DList as DL
 import Data.Default (Default, def)
 import Data.Foldable (fold, toList)
-import Data.HashMap.Lazy (HashMap)
 import qualified Data.HashMap.Lazy as HashMap
 import Data.List (intersperse)
 import Error.Diagnose.Report (Report)
-import Error.Diagnose.Report.Internal (prettyReport)
+import Error.Diagnose.Report.Internal (FileMap, prettyReport)
 import Error.Diagnose.Style (Annotation, Style)
 import Prettyprinter (Doc, Pretty, hardline, unAnnotate)
 import Prettyprinter.Render.Terminal (hPutDoc)
@@ -45,7 +45,7 @@ data Diagnostic msg
       -- ^ All the reports contained in a diagnostic.
       --
       --   Reports are output one by one, without connections in between.
-      (HashMap FilePath [String])
+      !FileMap
       -- ^ A map associating files with their content as lists of lines.
 
 instance Default (Diagnostic msg) where
@@ -57,7 +57,7 @@ instance Semigroup (Diagnostic msg) where
 #ifdef USE_AESON
 instance ToJSON msg => ToJSON (Diagnostic msg) where
   toJSON (Diagnostic reports files) =
-    object [ "files" .= fmap toJSONFile (HashMap.toList files)
+    object [ "files" .= fmap toJSONFile (fmap toList <$> (HashMap.toList files))
            , "reports" .= reports
            ]
     where
@@ -120,7 +120,10 @@ addFile ::
   String ->
   Diagnostic msg
 addFile (Diagnostic reports files) path content =
-  Diagnostic reports (HashMap.insert path (lines content) files)
+  let fileLines = lines content
+      lineCount = length fileLines
+      lineArray = listArray (0, lineCount - 1) fileLines
+   in Diagnostic reports (HashMap.insert path lineArray files)
 {-# INLINE addFile #-}
 
 -- | Inserts a new report into a diagnostic.