While looking at the code (I'm considering using this very nicely looking library for better error messages in [JaSpa/AlgST](https://github.com/JaSpa/AlgST)) I saw some potential for performance improvements. Some numbers:

|                      |    big/string | test-suite/string |
|---------------------:|--------------:|------------------:|
|           **v2.1.1** |   `284.50 ms` |        `778.9 μs` |
|         **improved** |    `58.71 ms` |        `686.5 μs` |
| absolute improvement | `- 225.79 ms` |       `- 92.4 μs` |
| relative improvement |  `- 79.36 % ` |       `- 11.86 %` |

These numbers show
1. there is large potential for improvement in large cases
2. smaller cases improve as well/don't get slower

Below I detail the benchmark setup I used and explain the changes I made. If you have any questions, please don't hesitate!

I verified that the test-suite output before and after is byte identical. (There *is* actually one change but the representation is the same: in the unicode.txt test case a sequence of spaces changes from each space being highlighted by itself to the whole chunk being highlighted at once. I don't really see where I introduced this but I won't investigate further as I think this change is harmless.)

### Benchmark Setup

The code is available at https://gist.github.com/JaSpa/0d6a267351c6d91daddb73838270ea72.  
Full results are at https://gist.github.com/JaSpa/032501608da016b84a6cb3a55377e723

The test case *big/string* generates a report with a single marker for every line of a file with ~3600 lines (6 concatenations of Report/Internal.hs at 73c2f2b—the specific size is admittedly quite arbitrary. The diagnostic is rendered using [`layoutCompact`](https://hackage.haskell.org/package/prettyprinter-1.7.1/docs/Prettyprinter-Internal.html#v:layoutCompact) to cut down on the influence of the layout algorithm.

The test case *test-suite/string* uses the same diagnostic/reports/files as the test suite and the same layouting as described above.

Additionally, I benchmarked the *big* case, as well as the *test-suite* case, with JSON output. These numbers change only very little. This is expected, as most of the changes revolve around `getLine_` which is not used for JSON output. I included it mostly for sanity checking that I don't something unrelated worse.

If you're interested I can help with integrating the benchmarks into the repository.

### Applied Improvements/Rationales

* Use difference lists when accumulating reports. This results in `addReport` running in *O(1)* instead of *O(n)*. Overall for *n* reports this improves the run time from *O(n²)* to *O(n)*.

  ```
  big/string		-132.3ms
  test-suite/string	  -1.9μs
  ```

* Calculate character widths per line in an array.

  Using an unboxed array has better performance than using a
  `HashMap Int Int`; although both data structures support O(1) access,
  the array has a better constant factor.

  Additionally, an array is suited very well since we we have a character
  width for every index i in [1..n].


  ```
  big/string		  -3.7ms
  test-suite/string	 -97.8μs
  ```

  Here the impact on the *test-suite/string* case is larger than the *big/string* case because in the former there are multiple markers and therefore the access performance carries more weight.

* Store a file's lines in an array. Another *O(n²)* to *O(n)* improvement (here *n* being the number of lines in a file): Haskell's lists are very not-suited for indexed access (`List.safeIndex`). `Seq` from *containers* is a better choice but since we do not have to add/remove any elements we can use an array which is even faster.
  
  ```
  big/string		-89.79ms
  test-suite/string	 +7.3μs
  ```

  The impact of this change is especially high in the *big/string* case because of the big file. The test-suite files, however, are very small. Here the `List.safeIndex` is nearly *O(1)* anyways and the array constructions constant factor results in the slight increases in runtime.