@Mesabloo 2022-07-02
Thank you very much for this very detailed and nicely formatted report.

I initially did not think too much about performance when creating this library. My thought process was that in most cases, this is an only once done task, which makes it way less performance critical.
However, such performance improvements cannot be ignored.

---------------

There are several issues with my current test suite, the primary one being that it is *not* a correct test suite (I only use it to check manually if the reports correctly constructed; ideally there would be a byte comparison to check that every single one is correct).
> If you're interested I can help with integrating the benchmarks into the repository.
I'd be happy to also integrate your tests! More tests is better than none.

-------------

As for the fixes:
- > Use difference lists when accumulating reports.

  I actually did not know about this package. Another thought I had was just to build a `Seq`, not sure in what way it would differ.
  This seems to have a very big impact (which is to be expected in fact, as my original code did `reports <> [newReport]` which is a big performance hole given the nature of haskell lists).
- > Calculate character widths per line in an array.

  Now I feel dumb about this one. I'm not sure why I chose `HashMap Int Int`, because even the indices are contiguous.
- > Store a file's lines in an array.

  I did not think about this one, but it's actually pretty nice (more so because the file never changes dynamically, it is a one time add to the diagnostic).
  
----------

> There is actually one change but the representation is the same: in the unicode.txt test case a sequence of spaces changes from each space being highlighted by itself to the whole chunk being highlighted at once. I don't really see where I introduced this but I won't investigate further as I think this change is harmless.

This is fine. The chunks you refer to are generated for the TAB character. So this is harmless because the whole chunk is colored anyway. Doing so on every space or for the whole chunk at once are both the same behavior (the second one might actually be slightly more efficient when rendering on the terminal?).