@JaSpa 2022-07-02
> Another thought I had was just to build a Seq, not sure in what way it would differ.

I did two additional checks: once using a `Seq` and once accumulating the reports in reverse order in a simple list and then throwing in a `reverse` before rendering.

The `Seq` version is *a lot* slower whereas the `reverse` version is nearly as good as the `DList` version in the big test case but still an increase in runtime over no change in the smaller case. (Absolute changes always in reference to v2.1.1.)

**`DList` version**
```
big/string		-132.3ms
test-suite/string	  -1.9μs
```

**`Seq` version**
```
big/string		 -66.6ms
test-suite/string	 +95.3μs (!)
```

**`reverse` version**
```
big/string		-128.5ms
test-suite/string	 +12.6μs (!)
```