@Mesabloo 2022-07-02
That's actually interesting that the `Seq` approach is that much slower. Thanks!