@Mesabloo 2022-07-03
I'm willing to merge this all into the master branch. Is there any more that you would like to try or change?
If not, I'll just merge this. :)